backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
numpy>=1.26.0
numpy-financial>=1.0.0
openpyxl>=3.1.0
anthropic>=0.40.0
python-dotenv>=1.0.0
cachetools>=5.5.0
pymupdf>=1.24.0
